+++ CHANGELOG.rst
@@ -2,8 +2,11 @@
  Changes for bioy
 ==================
 
-1.11-dev
-=====
+=======
+1.11.dev
+========
+ * ``bioy --split-condensed-assignments`` to retain assignment threshold assignments regardless of common 
+   condense ids assignments (GH 53)
  * `bioy classifier --include-ref-rank $RANK` will pull in ${RANK}_id and ${RANK}_name columns from taxonomy into details.
 
 1.11